sptool: Port lotus-miner actor withdraw
