+++ cli/spcli/util.go
@@ -1,0 +1,9 @@
+package spcli
+
+import (
+	"github.com/urfave/cli/v2"
+
+	"github.com/filecoin-project/go-address"
+)
+
+type ActorAddressGetter func(cctx *cli.Context) (address address.Address, err error)
